Refactor, use trait for common logic
