+++ src/Model/Resolver/CartResolver.php
@@ -53,6 +53,15 @@ abstract class CartResolver implements ResolverInterface
     }
 
     /**
+     * @param array|null $args
+     * @return bool
+     */
+    protected function isGuest(array $args = null): bool
+    {
+        return array_key_exists('guestCartId', $args);
+    }
+
+    /**
      * @param string|null $guestCartId
      * @return string
      * @throws NotFoundException
@@ -65,61 +74,17 @@ abstract class CartResolver implements ResolverInterface
             : $this->getCartIdForLoggedInUser();
     }
 
-    protected function CreatePaymentDetails(PaymentDetailsInterface $paymentInformation){
-        $rawTotals = $paymentInformation->getTotals();
-
-        return [
-            'payment_methods' => array_map(
-                function ($payment) {
-                    /** @var PaymentMethodInterface $payment */
-                    return [
-                        'code' => $payment->getCode(),
-                        'title' => $payment->getTitle(),
-                    ];
-                },
-                $paymentInformation->getPaymentMethods()
-            ),
-            'totals' => array_merge(
-                $rawTotals->getData(),
-                [
-                    'items_qty' => $rawTotals->getItemsQty(),
-                    'items' => array_map(function ($item) {
-                        /** @var TotalsItemInterface $item */
-                        return [
-                            TotalsItemInterface::KEY_ITEM_ID => $item->getItemId(),
-                            TotalsItemInterface::KEY_PRICE => $item->getPrice(),
-                            TotalsItemInterface::KEY_BASE_PRICE => $item->getBasePrice(),
-                            TotalsItemInterface::KEY_QTY => $item->getQty(),
-                            TotalsItemInterface::KEY_ROW_TOTAL => $item->getRowTotal(),
-                            TotalsItemInterface::KEY_BASE_ROW_TOTAL => $item->getBaseRowTotal(),
-                            TotalsItemInterface::KEY_ROW_TOTAL_WITH_DISCOUNT => $item->getRowTotalWithDiscount(),
-                            TotalsItemInterface::KEY_TAX_AMOUNT => $item->getTaxAmount(),
-                            TotalsItemInterface::KEY_BASE_TAX_AMOUNT => $item->getBaseTaxAmount(),
-                            TotalsItemInterface::KEY_TAX_PERCENT => $item->getTaxPercent(),
-                            TotalsItemInterface::KEY_DISCOUNT_AMOUNT => $item->getDiscountAmount(),
-                            TotalsItemInterface::KEY_BASE_DISCOUNT_AMOUNT => $item->getBaseDiscountAmount(),
-                            TotalsItemInterface::KEY_DISCOUNT_PERCENT => $item->getDiscountPercent(),
-                            TotalsItemInterface::KEY_PRICE_INCL_TAX => $item->getPriceInclTax(),
-                            TotalsItemInterface::KEY_BASE_PRICE_INCL_TAX => $item->getBasePriceInclTax(),
-                            TotalsItemInterface::KEY_ROW_TOTAL_INCL_TAX => $item->getRowTotalInclTax(),
-                            TotalsItemInterface::KEY_BASE_ROW_TOTAL_INCL_TAX => $item->getBaseRowTotalInclTax(),
-                            TotalsItemInterface::KEY_OPTIONS => $item->getOptions(),
-                            TotalsItemInterface::KEY_WEEE_TAX_APPLIED_AMOUNT => $item->getWeeeTaxAppliedAmount(),
-                            TotalsItemInterface::KEY_WEEE_TAX_APPLIED => $item->getWeeeTaxApplied(),
-                            TotalsItemInterface::KEY_NAME => $item->getName(),
-                        ];
-                    }, $rawTotals->getItems())]
-            )
-        ];
-    }
-
     /**
-     * @param array|null $args
-     * @return bool
+     * @return string
+     * @throws UnexpectedValueException
      */
-    private function isGuest(array $args = null): bool
+    private function getCartIdForLoggedInUser(): string
     {
-        return array_key_exists('guestCartId', $args);
+        try {
+            return $this->overriderCustomerId->getOverriddenValue();
+        } catch (NoSuchEntityException $e) {
+            throw new \UnexpectedValueException(__("Unable to retrieve cart id. guestCartId is missing or you are not logged in"), 13, $e);
+        }
     }
 
     /**
@@ -143,20 +108,6 @@ abstract class CartResolver implements ResolverInterface
     private function getCartForLoggedInUser()
     {
         try {
-            return $this->quoteManagement->getCartForCustomer($this->getCartIdForLoggedInUser());
-        } catch (NoSuchEntityException $e) {
-            throw new \UnexpectedValueException(__("Unable to retrieve cart. guestCartId is missing or you are not logged in"), 13, $e);
-        }
-    }
-
-    /**
-     * @return string
-     * @throws NoSuchEntityException
-     */
-    private function getCartIdForLoggedInUser(): string
-    {
-        try {
-            return $this->overriderCustomerId->getOverriddenValue();
             return $this->quoteManagement->getCartForCustomer($this->getCartIdForLoggedInUser());
         } catch (NoSuchEntityException $e) {
             throw new \UnexpectedValueException(__("Unable to retrieve cart. guestCartId is missing or you are not logged in"), 13, $e);
